1. Implement all atomic operations on POSIX platform 2. Fix mouse button event errors on X11 3. Fix wchar_t error on POSIX system: wchar_t on POSIX system is UTF32, instead of UTF16 4. Fix ASCII font class crashing on POSIX system

+++ src/priv/util/util/asciifont.cpp
@@ -31,7 +31,7 @@ public:
 
     virtual bool loadFontImage( FontImage & result, wchar_t ch )
     {
-        if( ch > 255 )
+        if( (unsigned int)ch > 255 )
         {
             GN_ERROR(sLogger)( "this function supports ASCII characters only." );
             return false;
